Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/browsing-context-names/choose-_blank-003-expected.txt


FAIL Context created by link targeting "_blank" should retain opener reference assert_equals: expected "true" but got "false"
Link
